refactor: extract alias methods to package-aliases.js
Move magento/* alias metapackage generation into its own module for
better separation of concerns. The alias feature bridges third-party
extension dependencies to mage-os/* packages.

- Create src/package-aliases.js with alias generation logic
- Export writePackage from package-modules.js for use by alias module
- Pass packageModules as dependency to avoid circular imports
